backend/checkout/serializers/product_serializers.py
from rest_framework import serializers
from checkout.models.product import Product

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ['id', 'name', 'code', 'price']
